requirements.txt
-e .
ruff
coverage
django_coverage_plugin
